+++ lib/knapsack/distributors/leftover_distributor.rb
@@ -1,14 +1,8 @@
 module Knapsack
   module Distributors
     class LeftoverDistributor < BaseDistributor
-      attr_reader :spec_pattern
-
       def report_specs
-        @report_specs ||= @report.keys
-      end
-
-      def all_specs
-        @all_specs ||= Dir[spec_pattern]
+        @report_specs ||= report.keys
       end
 
       def leftover_specs
@@ -16,14 +10,6 @@ module Knapsack
       end
 
       private
-
-      def post_initialize(args={})
-        @spec_pattern = args[:spec_pattern] || default_spec_pattern
-      end
-
-      def default_spec_pattern
-        'spec/**/*_spec.rb'
-      end
 
       def post_assign_spec_files_to_node
         leftover_specs.each do |spec_file|